refactor(cli): split tool argument parser
